requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
python-dotenv==1.0.1
requests==2.31.0
SQLAlchemy==2.0.28
typing-extensions==4.9.0
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
urllib3==2.2.1
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6 